+++ mem/mem.h
@@ -1,10 +1,10 @@
 struct mem {
-    unsigned long size;
-    unsigned long cap;
-    unsigned long data_size;
+    int size;
+    int cap;
+    int data_size;
     void *buf;
 };
 
-int meminit(struct mem *m, unsigned long data_size, unsigned long cap);
+int meminit(struct mem *m, int data_size, int cap);
 int memgrow(struct mem *m);
 void memfree(struct mem *m);
